[FEAT] DailyWorker 자정 마다 섭취한 물 저장 추가
